В баш-скрипты для работы с докером добавлена проверка его установки

+++ start.sh
@@ -1,8 +1,10 @@
 #!/bin/sh
 
+command -v docker >/dev/null 2>&1 || { echo "Докер не установлен. Установка: http://docs.docker.com/linux/started"; exit 1; }
+
 echo "Удаление старого контейнера ..."
 docker stop kingdom > /dev/null 2>&1
 docker rm kingdom > /dev/null 2>&1
 echo "Создание нового контейнера ..."
-docker run -d --name="kingdom" -v $(pwd):/kingdom rottenwood/kingdom kingdom/app/docker/init.sh
+docker run -d --name="kingdom" -v $(pwd):/kingdom rottenwood/kingdom
 echo "Контейнер создан!"
